generate_solvated_ensemble Bug fix
Problem occurs if solute molecule is the same as the solvent
Where atoms seem to be misidentified e.g. with H2O. Central
'solute' H2O looks like the identity of the oxygen has been
swapped with one of the hygrogens.

Patched so that if solute and solvent are the same we call
solvate with only one molecular species